+++ src/react/internal/TwoslashHover.tsx
@@ -5,10 +5,9 @@ import { TwoslashHover as TwoslashHover_client } from './TwoslashHover.client.js
 export function TwoslashHover(props: TwoslashHover.Props) {
   const { className = '', children } = props
   const [trigger, ...content] = React.Children.toArray(children)
-  const open = className?.includes('twoslash-query-persisted')
   if (!content) return null
   return (
-    <TwoslashHover_client className={className} open={open} trigger={trigger}>
+    <TwoslashHover_client className={className} trigger={trigger}>
       {content}
     </TwoslashHover_client>
   )