+++ src/waku/vite.ts
@@ -35,6 +35,7 @@ export async function vocs(options: vocs.Options = {}): Promise<PluginOption[]> 
 
   return [
     vocs_core(),
+    Plugins.rsdwNoExternal(),
     Plugins.allowServer(),
     PluginRsc({
       serverHandler: false,
@@ -50,7 +51,6 @@ export async function vocs(options: vocs.Options = {}): Promise<PluginOption[]> 
     Plugins.adapterAlias(wakuConfig),
     Plugins.notFound(),
     Plugins.patchRsdw(),
-    Plugins.patchRouterPrefetch(),
     Plugins.buildMetadata(wakuConfig),
     Plugins.staticBuild(wakuConfig),
     Plugins.privateDir(wakuConfig),